Testing: add Jest and Babel.
All tests will be included in the "tests" folder.
With the current state of Jest, the TypeScript file will not be compiled
into the dist/ folder, but will only be transpiled with babel; given
this fact, it is **highly** recommended to run `tsc --noEmit` before
running tests. This makes sure that no syntattic or semantic errors
occur before testing.
At the moment moving to ts-jest doesn't seem worth the effort.