API: add /request_password_reset endpoint
This endpoint requires no authentication whatsoever, because no disruptive action
is taken: it checks for a user; if such user exists, it adds a key to a database
and sends an email to the user with the key, otherwise, no action is taken.

Please note that the API still sends the exact same message for both outcomes,
this is intended, as a malicious user could have insights on emails existing or not
if we sent different messages.